+++ schemars/src/json_schema_impls/serdejson.rs
@@ -2,6 +2,7 @@ use crate::gen::SchemaGenerator;
 use crate::schema::*;
 use crate::JsonSchema;
 use serde_json::{Map, Number, Value};
+use std::borrow::Cow;
 use std::collections::BTreeMap;
 
 impl JsonSchema for Value {
@@ -9,6 +10,10 @@ impl JsonSchema for Value {
 
     fn schema_name() -> String {
         "AnyValue".to_owned()
+    }
+
+    fn schema_id() -> Cow<'static, str> {
+        Cow::Borrowed("AnyValue")
     }
 
     fn json_schema(_: &mut SchemaGenerator) -> Schema {
@@ -25,6 +30,10 @@ impl JsonSchema for Number {
         "Number".to_owned()
     }
 
+    fn schema_id() -> Cow<'static, str> {
+        Cow::Borrowed("Number")
+    }
+
     fn json_schema(_: &mut SchemaGenerator) -> Schema {
         SchemaObject {
             instance_type: Some(InstanceType::Number.into()),
@@ -33,3 +42,6 @@ impl JsonSchema for Number {
         .into()
     }
 }
+
+#[cfg(feature = "raw_value")]
+forward_impl!(serde_json::value::RawValue => Value);